refactor: structures
