+++ gatsby-browser.tsx
@@ -7,7 +7,7 @@
 // You can delete this file if you're not using it
 
 import mediumZoom from "medium-zoom";
-import { Sign } from "./src/components/Sign.tsx";
+import { Sign } from "./src/utils/Sign.tsx";
 // import { render } from "react-dom";
 // import { loadableReady } from "@loadable/component";
