+++ src/components/ClientSide.tsx
@@ -1,0 +1,19 @@
+/************************
+ * Made by [MR Ferry™]  *
+ * on Mei 2024          *
+ ************************/
+
+import React, { useEffect, useState } from "react";
+
+export function ClientSide({ children }: React.PropsWithChildren) {
+  const [onClient, setOnClient] = useState(false);
+
+  useEffect(() => {
+    setOnClient(true);
+  }, []);
+
+  if (onClient) {
+    return <>{children}</>;
+  }
+  return <></>;
+}
