+++ src/components/CustomPageContainer.tsx
@@ -5,11 +5,11 @@
 
 import Layout from "./Layout";
 import React from "react";
-import { getPostTags, getPublishDateTime } from "../utils/GatsbyanUtils";
+import { getPublishDateTime } from "../utils/GatsbyanUtils";
 import { CustomPageProp } from "../types/DataTypes";
 import { Slice } from "gatsby";
 
-const CustomPage = ({ children, site, customPost }: React.PropsWithChildren<CustomPageProp>) => {
+const CustomPageContainer = ({ children, site, customPost }: React.PropsWithChildren<CustomPageProp>) => {
   const repo = site.repo;
   return (
     <Layout>
@@ -18,7 +18,6 @@ const CustomPage = ({ children, site, customPost }: React.PropsWithChildren<Cust
           <div className="title posted">{customPost.title}</div>
           <div className="title text-info mb-2">
             <span className="page-info">{getPublishDateTime(customPost.publishDate)}</span>
-            <span className="page-info">{getPostTags(null)}</span>
           </div>
           <div className="post-container pt-0 content-post">{children}</div>
           <Slice alias="Comment" repo={repo} />
@@ -28,4 +27,4 @@ const CustomPage = ({ children, site, customPost }: React.PropsWithChildren<Cust
   );
 };
 
-export default CustomPage;
+export default CustomPageContainer;
