+++ src/components/sidebar/Tags.tsx
@@ -1,18 +1,16 @@
 import React from "react";
-import { graphql, useStaticQuery } from "gatsby";
-import { getTags } from "../../utils/GatsbyanUtils";
+import { graphql, Link, useStaticQuery } from "gatsby";
+import { kebabCase } from "../../utils/GatsbyanUtils";
 import { TagsData } from "../../types/DataTypes";
 
 const Tags = () => {
-  const { allContentfulBlogPost }: TagsData = useStaticQuery(
-    graphql`
-      query Tags {
-        allContentfulBlogPost {
-          tags: distinct(field: { tags: SELECT })
-        }
+  const { allContentfulBlogPost }: TagsData = useStaticQuery(graphql`
+    query Tags {
+      allContentfulBlogPost {
+        tags: distinct(field: { tags: SELECT })
       }
-    `
-  );
+    }
+  `);
 
   return (
     <>
@@ -20,7 +18,7 @@ const Tags = () => {
       <div className="d-block">
         {allContentfulBlogPost.tags.map((tag, i) => (
           <div key={i} className="d-inline-block p-1 tag-link">
-            {getTags(tag)}
+            <Link to={`/tags/${kebabCase(tag)}`}>{tag}</Link>
           </div>
         ))}
       </div>
