+++ src/utils/Photo.tsx
@@ -1,5 +1,10 @@
 /************************
  * Made by [MR Ferry™]  *
+ * on Mei 2024          *
+ ************************/
+
+/************************
+ * Made by [MR Ferry™]  *
  * on April 2020        *
  ************************/
