+++ src/utils/SiteMetadata.tsx
@@ -1,5 +1,10 @@
 /************************
  * Made by [MR Ferry™]  *
+ * on Mei 2024          *
+ ************************/
+
+/************************
+ * Made by [MR Ferry™]  *
  * on May 2021          *
  ************************/
